ABC SMC with single cell sample trees using tree Branching Times as summary statistics

Performs ABC SMC simulations to recover a given parameter set for a tumour simulation using its single cell sampling phylogenetic trees.

ABCSMCwithTreeSamplesBT(
  sim.id = 1,
  grid = 50,
  mu.rl = 10,
  s.rl = 2,
  t.rl = 3,
  d.rl = 0,
  a.rl = 1,
  random.state = 123,
  nsamples = 10,
  dimension = "2D",
  output.dir = "data/inferred_params/",
  sampling = "tree",
  params.to.rec = "mu_t_s_d_a",
  sigma = 0.5,
  N = 100,
  rounds = 10,
  ncores = 2,
  params = data.table(name = c("mu", "t", "s", "d", "a"), min = c(1, 1, 1, 0, 0), max =
    c(100, 20, 5, 1, 1), fix = c(F, F, F, F, F)),
  target.tree = NULL,
  target.popsize = NULL
)

Arguments

sim.id
Simulated tumour ID.

grid
Tumour space grid size.

mu.rl
Mutation rate.

s.rl
Selective advantage.

t.rl
Time to a new subclone introduction.

d.rl
Death rate.

a.rl
Aggression (cell pushing probability scaled by its location on the grid).

random.state
Random state.

nsamples
Number of single cells to sample.

dimension
Tumour space grid dimension.

output.dir
Path to a directory where the output files will be exported.

sampling
Sampling scheme.

params.to.rec
String of parameter names (will be set as a directory name where the inferred particle files will be saved per SMC round).

sigma
Permutation kernel scale factor.

N
Number of simulations to run per SMC round.

rounds
Number of SMC rounds.

ncores
Number of cores to run in parallel with.

params
Data.table containing parameter names, uniform prior distribution limits and flags to fix each parameter at given (previously inferred or already known) values or not. For exact format of the table, please see the examples section.

target.tree
Phylogenetic tree (phylo object) to recover the parameters for.

target.popsize
Estimated size (as number of cells) of the tumour to infer the parameters for.

Value

Exports inferred set of parameters by each SMC round (as .csv files) into a given directory.